requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.0
django-toolbelt==0.0.1
heroku==0.1.4
honcho==0.5.0
python-dateutil==1.5
requests==2.4.1
static3==0.5.1

python-social-auth
